requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.20.0
requests==2.31.0
python-dotenv==1.0.1
stripe==8.5.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.124.0
httpx==0.27.0
pydantic==2.6.4
